[template] Adding back a case that needed coverage.
Since a SECTION only executes when its value is truth-y, the OR clause
will only execute when the value is false-y (false, "", 0, etc).  Since
SECTIONs do not look up the stack to resolve variables, embedding a
nested block inside an OR rarely makes sense.  However, since it is
possible we must have a contrived test case to cover that case.