[template] Compiler in validation mode by default.
* Put the new JSONT compiler into validation mode by default. This
  ensures that all errors will be captured and no exceptions thrown
  for invalid templates.
* If any errors occur, they are appended to the output inside of an HTML
  comment block.
* Dropped the NodeJsContext class in favor of new TemplateInfo and
  TemplateType classes.  A slightly cleaner and more flexible way of
  passing the compiler information about the source template.

Note: in validation mode no errors are thrown, and invalid instructions
are either converted to TEXT or a safe alternative to the original
instruction.  This will often not result in the intended output, but the
result will be executable and at least produce enough output to carry
the error messages to the browser.