[template] Child errors, safe execution mode.
Now errors can have children.  If a template is executing and applies a
partial which in turn has syntax errors, this will result in one or more
nested errors.

Added a "safe execution" mode where all exceptions are suppressed and
replaced with a list of errors.  All errors that occur during
compilation or execution will now be appended to the template output.