[template] Flatten POJOs instead of going to/from String.
Before passing the JsonNode to the compiler, recursively eliminate all
POJONodes and replace them with their mapped type.

This is still not the ideal solution, and I'm working out a way to
flatten the POJOs to JsonNode in a single pass by creating a custom
mapper.  However, its much faster than going to/from String to get rid
of the POJONodes.