[template] Improve error handling in APPLY plugin.
In safe mode, no exceptions should be raised.  Errors in syntax and
execution will be appended to the context, so they can be output as
desired by the client of the compiler.  Unexpected exceptions will
be appended to the context and logged.

In non-safe mode, all errors will immediately raise an exception.

Added a test to ensure unexpected exceptions (NPE, etc.) are sent to the
logging hook, if one exists.   These are execeptions thrown by the
framework, not due to errors in the template.  Also added an NPE
formatter plugin to support testing unexpected exception-handling.