[template] Added BINDVAR instruction {.var} and tests.
New instruction BINDVAR adds the ability to bind a name to a value at the
currently-executing scope.  This binding will be visible to all
instructions after the BINDVAR and to all nested scopes lower than the
scope in which the binding is made.

The syntax supports drilling into @variables.

    {.var @person person}
    {@person.name} is {@person.age}.

It can simplify some situations where you're iterating over 2 nested
arrays of near-identical objects:

    {.repeated section managers}
        {.var @boss @}
        {.var @boss-idx @index}
        {.repeated section employees}
            {@boss-idx}.{@index} - {name} is managed by {@boss.name}
        {.end}
        {.alternates with}---{.end}
    {.end}

Could produce this:

  1.1 - Peter is managed by Bill
  1.2 - Michael is managed by Bill
  ---
  2.1 - Samir is managed by Bob
  ...