[template] Added ReferenceScanner and related changes.
The ReferenceScanner will traverse a compiled template instruction tree,
collecting stats on how many references it finds for instruction types,
predicates and formatters.  It also records the nested tree of variable
references, to show which parts of the JSON context a template
references.

Also added a method to let predicates indicate which of their arguments
are variable references, so these references can be recorded by the
ReferenceScanner.