Ensure alternates-with has access to correct context.
Since alternates-with blocks execute between the elements of the
parent block's array, it is ambiguous which context the block should
see:  parent, current element or next element.

I've changed this to expose the current element's scope to the
alternates-with block.  So as a loop executes from i = 0 to N,
the alternates-with block will see the scope of iteration i.