Bindvar now supports formatters, Formatter.apply() changes
Now formatters apply() method takes a new argument, the JsonNode they
will operate on, and return the modified JsonNode.  This allows
the BindVar instruction to apply a formatter to a referenced node
and create a binding for the formatted value, all without modifying
the current node pointed to by the context.