Instruction soft/hard limiting.
This allows soft/hard instruction limits to be set, monitored and
enforced.  With this API you could monitor the number of instructions
executed and emit a log statement when a soft limit is exceeded, or
throw an error when the hard limit is exceeded.  It also provides a way
to count the number of instructions executed, to provide a measurement
of template complexity at runtime.