Allow variables starting with '$', change multi-var delimiter.
Template variables can now be '$'-prefixed.

The '$' was previously used to delimit multi-variable syntax, to reduce the
chance of false matches when parsing inlined JavaScript. The delimiter is now
comma and the parse is more strict, to only allow multi-variable syntax when
pipe-ing the vars into a formatter.

Also reduced false positives generally where JS boolean operators would attempt
to parse as a formatter name and fail.