Add a pre-processor scope to the template syntax.
Adding this as there are cases where a template needs to be
pre-processed using the template language itself.

Instructions are placed into pre-processor scope by prefixing them with
the carat '^' character, implying the instruction is "lifted up" into
the pre-processor.  When the compiler is in pre-processor mode, only
scoped instructions will be evaluated.